chore: rename to `CommonInputs`
